+++ changelog.md
@@ -43,3 +43,4 @@
  * Fix Large Distillery Structure: adds tiered casing support and blocks multi-fluid hatches (#5)
  * Internal: Moved to non-deprecated mod loaded check (#9)
  * Adds compat for GTCEu V2.1.5-beta
+ * Removed muffler hatch from the Large Laser Engraver